display_screen.dart: Moved photo display screen to new file
